Prep to use local maven repository
These changes are in preparation for the netcdf-java build on travis-ci
to test PRs against the TDS master branch. To do so, we need to tell the
TDS build to look in the local maven repository for snapshots. Normally,
this is not something you'd want to do unless you were testing
netcdf-java changes against the TDS locally, therefore these are
commended out. The netcdf-java travis script will use sed to take care of
uncommenting the two entries.