Fix test expected files
The move from using joda-time to CalendarDate changed the way date
time strings are formatted. Specifically, for a CalendarDate, the
toString() method will return times down to the millisecond only if
milliseconds does not equal zero. This commit changes the expected
files on disk to reflect that change. No information was actually
lost by the string format change, so I decided to update the test
expected files.