Make NcML compact before adding to dataset tracker
DatasetTrackerChronicle does not work when the NcML found in a catalog
become too large for the read() method of ObjectInput to fully consume.
Chaing the read() method to readFully() does not help, as there appears
to be other issues related to the non-blocking design of ChronicleMap.

While this does not fix the issue of an arbitrairly large block of
NcML not working with ChronicleMap, it does reduce the size of what is
trying to be stored by removing unnecessary whitespace, and stripping
out all comments. This might help mitigate the issue reported in
https://github.com/Unidata/tds/issues/56.