Allow TDS to use both old and new netCDF-Java API
This PR enables the TDS and its tests to configure whether or not the
new netCDF-Java API is used when reading data (that is, whether or not
it will use the new NetcdfFiles and NetcdfDatasets methods). If the
dataset being served by the TDS is on an object store, it will be
accessed via the new API, as that is the only choice.