Correct time_coverage_start attr for point test
The time_coverage_start global attribute in NcML files used by the
DsgSubsetWriterTest do not have the correct value based on the data
defined in the file. The new point writer in netCDF-Java will update
the time_coverage_start/stop attrs based on what is actually written,
and so these values must be correct in the first place or the tests will
fail once Unidata/netcdf-java#319 is merged.