Enable use of wmsConfig.xml
During the move from ncWms 1.x to edal-java 2.x, use of the
wmsConfig.xml was not working and had to be reimplemented. Many of the
classes involved in making the custom defaults (by path, standard name,
or variable name) work was able to be reused, although porting to use
edal-java took a bit of time. A few new classes had to be created as
well.