Implement a generic disk-persisted cache
Create a generic, two-level disk persisted cache to replace the chronicle-map
based caches. The level 1 cache is an in-memory LRU cache implemented with
Guava. The level 2 cache holds key/values in an on-disk EclipseStore storage
layer. The level 1 cache is item limited; the level 2 cache is lazily indexed,
and references can be lazily loaded to feed the level 1 cache.